prog bookfav + details
mayor parte del back hecha, falta login y registro y ultimar detalles de seguridad

+++ Bybliotheca/src/main/java/pi/bybliotheca/controller/BookController.java
@@ -1,9 +1,14 @@
 package pi.bybliotheca.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import pi.bybliotheca.entity.Book;
+import pi.bybliotheca.entity.User;
+import pi.bybliotheca.repository.UserRepository;
 import pi.bybliotheca.service.BookService;
+import pi.bybliotheca.service.BorrowingService;
 
 import java.util.List;
 
@@ -13,24 +18,51 @@ public class BookController {
     @Autowired
     private BookService service;
 
+    @Autowired
+    UserRepository userRepository;
+
     @PostMapping("/addBook")
     public Book addBook(@RequestBody Book book){
-        return service.saveBook(book);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = userRepository.findByUsername(auth.getPrincipal().toString());
+        if(loggedUser.getRole().equals("ADMIN")) {
+            return service.saveBook(book);
+        } else {
+            throw new SecurityException("Invalid operation");
+        }
     }
 
     @PutMapping("/updateBook")
     public Book updateBook(@RequestBody Book book){
-        return service.updateBook(book);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = userRepository.findByUsername(auth.getPrincipal().toString());
+        if(loggedUser.getRole().equals("ADMIN")) {
+            return service.updateBook(book);
+        } else {
+            throw new SecurityException("Invalid operation");
+        }
     }
 
     @DeleteMapping("/deleteBook/{id}")
-    public String deleteBook(@PathVariable int id){
-        return service.deleteBook(id);
+    public void deleteBook(@PathVariable int id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = userRepository.findByUsername(auth.getPrincipal().toString());
+        if(loggedUser.getRole().equals("ADMIN")) {
+            service.deleteBook(id);
+        } else {
+            throw new SecurityException("Invalid operation");
+        }
     }
+
 
     @GetMapping("/books")
     public List<Book> getBooks(){
         return service.getBooks();
+    }
+
+    @GetMapping("/books/{id}")
+    public Book getBook(@PathVariable int id){
+        return service.getBookById(id);
     }
 
     @GetMapping("/books/{author}")
@@ -44,5 +76,7 @@ public class BookController {
     }
 
 
-
 }
+
+
+
